test(adapters): move openapi generated code to other package to exclude from unit tests
