+++ adapters/src/test/java/dev/molaya/tests/adapters/Stubs.java
@@ -1,6 +1,6 @@
 package dev.molaya.tests.adapters;
 
-import dev.molaya.tests.adapters.in.rest.dto.ProductDetail;
+import dev.molaya.tests.adapters.in.rest.gen.openapi.dto.ProductDetail;
 import dev.molaya.tests.domain.Product;
 import java.math.BigDecimal;
 import java.util.Set;
@@ -25,8 +25,8 @@ public final class Stubs {
         return detail;
     }
 
-    public static dev.molaya.tests.adapters.out.client.dto.ProductDetail productDetailOUT() {
-        final var detail = new dev.molaya.tests.adapters.out.client.dto.ProductDetail();
+    public static dev.molaya.tests.adapters.out.client.gen.openapi.dto.ProductDetail productDetailOUT() {
+        final var detail = new dev.molaya.tests.adapters.out.client.gen.openapi.dto.ProductDetail();
         detail.setId("A");
         detail.setName("Product 1");
         detail.setPrice(BigDecimal.valueOf(10.0));
